requirements.txt
absl-py==2.1.0
ace_tools==0.0
asttokens==2.4.1
axial_positional_embedding==0.2.1
backcall==0.2.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
cycler==0.12.1
Deprecated==1.2.14
docopt==0.6.2
einops==0.5.0
executing==2.1.0
fonttools==4.53.1
google-auth==2.32.0
google-auth-oauthlib==1.0.0
grpcio==1.65.1
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
joblib==1.4.2
kiwisolver==1.4.5
llvmlite==0.41.1
local-attention==1.4.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.0
matplotlib-inline==0.1.7
mkl-service==2.4.0
mpmath==1.3.0
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
packaging==24.1
pandas==1.5.3
parso==0.8.4
patool==1.12
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.4.13
prettytable==3.10.2
product_key_memory==0.1.10
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
reformer-pytorch==1.4.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.2
sktime==0.16.1
stack-data==0.6.3
sympy==1.11.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tqdm==4.64.1
traitlets==5.14.3
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
yarg==0.1.10
zipp==3.19.2
